rendermodel uint int comparison fix

+++ src/engine/render/rendermodel.cpp
@@ -682,7 +682,7 @@ void resetmodelbatches()
 void addbatchedmodel(model *m, batchedmodel &bm, int idx)
 {
     modelbatch *b = nullptr;
-    if(batches.size() > m->batch)
+    if(static_cast<int>(batches.size()) > m->batch)
     {
         b = &batches[m->batch];
         if(b->m == m && (b->flags & Model_Mapmodel) == (bm.flags & Model_Mapmodel))
